Debug & Adjust Dao & Ctrl & Update Doc

+++ app/controller/DocumentCtrl.py
@@ -2,6 +2,7 @@ import os
 
 from flask import Blueprint, g, request
 from flask_httpauth import HTTPTokenAuth
+from werkzeug.utils import secure_filename
 
 from app.config.Config import Config
 from app.database.DbStatusType import DbStatusType
@@ -49,9 +50,10 @@ def apply_blue(blue: Blueprint, auth: HTTPTokenAuth):
         """ 插入文档 (DB + FS) """
         try:
             upload_file = request.files.get('file')
-            req_filename = request.form['filename']
+            # req_filename = request.form['filename']
             req_docClass = int(request.form['doc_class_id'])
-            if not (upload_file and req_filename and req_docClass):
+            # if not (upload_file and req_filename and req_docClass):
+            if not (upload_file and req_docClass):
                 raise ParamError(ParamType.FORM)
         except:
             raise ParamError(ParamType.FORM)
@@ -65,7 +67,8 @@ def apply_blue(blue: Blueprint, auth: HTTPTokenAuth):
             return Result.error(ResultCode.SAVE_FILE_FAILED).setMessage('Save Document Failed').json_ret()
 
         # Database
-        document = Document(did=-1, filename=req_filename, uuid=server_filename, docClass=req_docClass)
+        filename = secure_filename(upload_file.filename)
+        document = Document(did=-1, filename=filename, uuid=server_filename, docClass=req_docClass)
         status, new_document = DocumentDao().insertDocument(uid=g.user, document=document)
         if status == DbStatusType.FOUNDED:  # -1 永远不会
             os.remove(os.path.join(server_filepath, server_filename))
@@ -84,6 +87,8 @@ def apply_blue(blue: Blueprint, auth: HTTPTokenAuth):
             req_id = int(request.form['id'])
             req_filename = request.form['filename']
             req_docClass = int(request.form['doc_class_id'])
+            if not req_filename or not (FileUtil.is_document(req_filename) or FileUtil.is_image(req_filename)):
+                raise ParamError(ParamType.FORM)
         except:
             raise ParamError(ParamType.FORM)
         req_doc = Document(did=req_id, filename=req_filename, docClass=req_docClass)
@@ -108,7 +113,8 @@ def apply_blue(blue: Blueprint, auth: HTTPTokenAuth):
             return Result.error(ResultCode.DATABASE_FAILED).setMessage("Document Delete Failed").json_ret()
         else:  # Success
             server_filepath = f'{Config.UPLOAD_DOC_FOLDER}/{g.user}/{document.uuid}'
-            os.remove(server_filepath)
+            if os.path.exists(server_filepath):
+                os.remove(server_filepath)
             return Result.ok().setData(document.to_json()).json_ret()
